TEST DE PRESTAMO: PASADO CON EXITO.

+++ Migrations/20200605042055_Migracion_Inicial.cs
@@ -17,7 +17,7 @@ namespace ProgramaPrestamos.Migrations
                     Apellido = table.Column<string>(nullable: true),
                     Cedula = table.Column<string>(nullable: true),
                     Historial = table.Column<string>(nullable: true),
-                    Balance = table.Column<int>(nullable: false)
+                    Balance = table.Column<double>(nullable: false)
                 },
                 constraints: table =>
                 {
